GLC
This objective of this assignment is to obtain a good understanding of 3D camera models, basic
3D transformations, and some geometric computations. You have one week and I recommend
you start immediately. You will write a program that implements a Generalized Linear Camera
(GLC) Model and will demonstrate a subset of the possible camera models it can imitate. You
program will make use of OpenGL but most of the work will be vector math and geometric
computations.